+++ onnxruntime/core/providers/tensorrt/tensorrt_execution_provider.cc
@@ -743,7 +743,7 @@ std::unique_ptr<IndexedSubGraph> TensorrtExecutionProvider::GetSubGraph(SubGraph
 
   // Generate unique kernel name for TRT subgraph
   HashValue model_hash = 0;
-  int id = TRTGenerateId(graph, model_hash);
+  int id = TRTGenerateModelId(graph, model_hash, trt_model_id_);
   std::string subgraph_id = std::to_string(model_hash) + "_" + std::to_string(id);
   auto meta_def = IndexedSubGraph_MetaDef::Create();
   const std::string graph_type = graph.IsSubgraph() ? "subgraph" : "graph";
@@ -2177,97 +2177,4 @@ void TensorrtExecutionProvider::RegisterStreamHandlers(IStreamCommandHandleRegis
   RegisterCudaStreamHandles(stream_handle_registry, OrtDevice::GPU, allocator, true, stream_, external_stream_, external_cudnn_handle_, external_cublas_handle_);
 }
 
-int TensorrtExecutionProvider::TRTGenerateId(const GraphViewer& graph_viewer, HashValue& model_hash) {
-  model_hash = 0;
-
-  // find the top level graph
-  const Graph* cur_graph = &graph_viewer.GetGraph();
-  while (cur_graph->IsSubgraph()) {
-    cur_graph = cur_graph->ParentGraph();
-  }
-
-  const Graph& main_graph = *cur_graph;
-  uint32_t hash[4] = {0, 0, 0, 0};
-
-  auto hash_str = [&hash](const std::string& str) {
-    MurmurHash3::x86_128(str.data(), gsl::narrow_cast<int32_t>(str.size()), hash[0], &hash);
-  };
-
-  // Use model name instead of path to avoid cache regeneration if path changes
-  const auto& model_path = main_graph.ModelPath();
-  if (!model_path.IsEmpty()) {
-    // Get model name
-    PathString path_string = model_path.GetComponents().back();
-    char arr[256];
-#ifdef _WIN32
-    wcstombs_s(nullptr, arr, sizeof(arr), path_string.c_str(), sizeof(arr));
-#else
-    strcpy(arr, path_string.c_str());
-#endif
-    std::string model_name(arr);
-    LOGS_DEFAULT(INFO) << "[TensorRT EP] Model name is " << model_name;
-    // Ensure enough characters are hashed in case model names are too short
-    int32_t model_name_length = gsl::narrow_cast<int32_t>(model_name.size());
-    constexpr int32_t hash_string_length = 500;
-    std::string repeat_model_name = model_name;
-    for (int i = model_name_length; i > 0 && i < hash_string_length; i += model_name_length) {
-      repeat_model_name += model_name;
-    }
-    hash_str(repeat_model_name);
-  } else {
-    LOGS_DEFAULT(INFO) << "[TensorRT EP] Model path is empty";
-  }
-
-  // fingerprint the main graph by hashing graph inputs
-  for (const auto* node_arg : main_graph.GetInputsIncludingInitializers()) {
-    hash_str(node_arg->Name());
-  }
-
-  // hashing output of each node
-  const int number_of_ort_nodes = graph_viewer.NumberOfNodes();
-  std::vector<size_t> nodes_vector(number_of_ort_nodes);
-  std::iota(std::begin(nodes_vector), std::end(nodes_vector), 0);
-  const std::vector<NodeIndex>& node_index = graph_viewer.GetNodesInTopologicalOrder();
-  for (const auto& index : nodes_vector) {
-    const auto& node = graph_viewer.GetNode(node_index[index]);
-    for (const auto* node_arg : node->OutputDefs()) {
-      if (node_arg->Exists()) {
-        hash_str(node_arg->Name());
-      }
-    }
-  }
-
-#ifdef __linux__
-  hash_str("LINUX");
-#elif defined(_WIN32)
-  hash_str("WINDOWS");
-#endif
-
-#ifdef ORT_VERSION
-  hash_str(ORT_VERSION);
-#endif
-
-#ifdef CUDA_VERSION
-  hash_str(std::to_string(CUDA_VERSION));
-#endif
-
-#if defined(NV_TENSORRT_MAJOR) && defined(NV_TENSORRT_MINOR)
-  std::string TRT_VERSION = std::to_string(NV_TENSORRT_MAJOR) + "." + std::to_string(NV_TENSORRT_MINOR);
-  hash_str(TRT_VERSION);
-#endif
-
-  model_hash = hash[0] | (uint64_t(hash[1]) << 32);
-
-  // return the current unique id, and increment to update
-  return trt_model_id_[model_hash]++;
-}
-
-// Call TRTGenerateModelId to generate hash id for TRT engine cache
-int TensorrtExecutionProvider::TRTGenerateModelId(const GraphViewer& graph_viewer, HashValue& model_hash) {
-  // if the EP is shared across multiple sessions there's a very small potential for concurrency issues.
-  // use a lock when generating an id to be paranoid
-  static OrtMutex mutex;
-  std::lock_guard<OrtMutex> lock(mutex);
-  return TRTGenerateId(graph_viewer, model_hash);
-}
 }  // namespace onnxruntime
